shift helper methods to different/new files to fix cop failures
